Make test fakers easier to reuse
